@@ -3,17 +3,20 @@ description: Sistem booking lapangan olahraga terintegrasi
 publish_to: 'none'
 version: 0.1.0
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^6.0.5
-  http: ^0.13.6
+  provider: ^6.0.0
+  http: ^1.2.0
   intl: ^0.18.0
   uuid: ^3.0.7
   ionicons: ^0.2.1
+  firebase_core: ^4.2.1
+  firebase_auth: ^6.1.2
+  cloud_firestore: ^6.1.0
 
 
 dev_dependencies:
