@@ -4,59 +4,185 @@ import '../../providers/auth_provider.dart';
 import '../../routes.dart';
 
 class LoginScreen extends StatefulWidget {
-  const LoginScreen({Key? key}) : super(key: key);
+  const LoginScreen({super.key});
 
   @override
   State<LoginScreen> createState() => _LoginScreenState();
 }
 
 class _LoginScreenState extends State<LoginScreen> {
+  // Controller Text Input
   final _emailCtrl = TextEditingController();
   final _passCtrl = TextEditingController();
-  bool _loading = false;
+  
+  bool _isLoading = false;
+  final _formKey = GlobalKey<FormState>();
+
+  @override
+  void dispose() {
+    _emailCtrl.dispose();
+    _passCtrl.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
-    final auth = Provider.of<AuthProvider>(context);
     return Scaffold(
-      appBar: AppBar(title: const Text('Login - Sportify')),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          children: [
-            TextField(
-                controller: _emailCtrl,
-                decoration: const InputDecoration(labelText: 'Email')),
-            TextField(
-                controller: _passCtrl,
-                obscureText: true,
-                decoration: const InputDecoration(labelText: 'Password')),
-            const SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: _loading
-                  ? null
-                  : () async {
-                      setState(() => _loading = true);
-                      final ok =
-                          await auth.login(_emailCtrl.text, _passCtrl.text);
-                      setState(() => _loading = false);
-                      if (ok) {
-                        Navigator.pushReplacementNamed(context, Routes.home);
-                      } else {
-                        ScaffoldMessenger.of(context).showSnackBar(
-                            const SnackBar(content: Text('Login gagal')));
-                      }
-                    },
-              child: _loading
-                  ? const CircularProgressIndicator()
-                  : const Text('Login'),
+      backgroundColor: Colors.white,
+      body: SafeArea(
+        child: Center(
+          child: SingleChildScrollView(
+            padding: const EdgeInsets.all(24.0),
+            child: Form(
+              key: _formKey,
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: [
+                  // --- HEADER ---
+                  const Icon(Icons.sports_soccer, size: 64, color: Colors.green),
+                  const SizedBox(height: 16),
+                  const Text(
+                    "Selamat Datang Kembali",
+                    textAlign: TextAlign.center,
+                    style: TextStyle(
+                      fontSize: 24,
+                      fontWeight: FontWeight.bold,
+                      color: Colors.black87,
+                    ),
+                  ),
+                  const SizedBox(height: 8),
+                  const Text(
+                    "Masuk untuk mulai booking lapangan",
+                    textAlign: TextAlign.center,
+                    style: TextStyle(color: Colors.grey),
+                  ),
+                  const SizedBox(height: 48),
+
+                  // --- INPUT EMAIL ---
+                  TextFormField(
+                    controller: _emailCtrl,
+                    keyboardType: TextInputType.emailAddress,
+                    decoration: _inputDecoration("Email", Icons.email_outlined),
+                    validator: (value) =>
+                        value!.isEmpty ? "Email tidak boleh kosong" : null,
+                  ),
+                  const SizedBox(height: 16),
+
+                  // --- INPUT PASSWORD ---
+                  TextFormField(
+                    controller: _passCtrl,
+                    obscureText: true,
+                    decoration: _inputDecoration("Password", Icons.lock_outline),
+                    validator: (value) =>
+                        value!.isEmpty ? "Password tidak boleh kosong" : null,
+                  ),
+                  const SizedBox(height: 24),
+
+                  // --- TOMBOL LOGIN ---
+                  ElevatedButton(
+                    onPressed: _isLoading ? null : _handleLogin,
+                    style: ElevatedButton.styleFrom(
+                      backgroundColor: Colors.green,
+                      foregroundColor: Colors.white,
+                      padding: const EdgeInsets.symmetric(vertical: 16),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(12),
+                      ),
+                      elevation: 0,
+                    ),
+                    child: _isLoading
+                        ? const SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: CircularProgressIndicator(
+                                color: Colors.white, strokeWidth: 2),
+                          )
+                        : const Text(
+                            "MASUK",
+                            style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                          ),
+                  ),
+
+                  const SizedBox(height: 24),
+
+                  // --- REGISTER LINK ---
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      const Text("Belum punya akun? ", style: TextStyle(color: Colors.grey)),
+                      GestureDetector(
+                        onTap: () {
+                          Navigator.pushNamed(context, Routes.register);
+                        },
+                        child: const Text(
+                          "Daftar Sekarang",
+                          style: TextStyle(
+                            color: Colors.green,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ],
+              ),
             ),
-            TextButton(
-                onPressed: () => Navigator.pushNamed(context, Routes.register),
-                child: const Text('Register'))
-          ],
+          ),
         ),
       ),
     );
+  }
+
+  // Desain Input Field
+  InputDecoration _inputDecoration(String label, IconData icon) {
+    return InputDecoration(
+      labelText: label,
+      prefixIcon: Icon(icon, color: Colors.grey),
+      border: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(12),
+        borderSide: BorderSide(color: Colors.grey.shade300),
+      ),
+      enabledBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(12),
+        borderSide: BorderSide(color: Colors.grey.shade300),
+      ),
+      focusedBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(12),
+        borderSide: const BorderSide(color: Colors.green, width: 2),
+      ),
+      filled: true,
+      fillColor: Colors.grey.shade50,
+      contentPadding: const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
+    );
+  }
+
+  // Logika Login
+  Future<void> _handleLogin() async {
+    // 1. Cek apakah form valid (tidak kosong)
+    if (!_formKey.currentState!.validate()) return;
+
+    // 2. Mulai Loading
+    setState(() => _isLoading = true);
+
+    // 3. Panggil Provider
+    final error = await context.read<AuthProvider>().login(
+      _emailCtrl.text.trim(), // Hapus spasi di awal/akhir
+      _passCtrl.text.trim(),
+    );
+
+    // 4. Stop Loading
+    if (mounted) setState(() => _isLoading = false);
+
+    // 5. Cek Hasil
+    if (error != null && mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text(error),
+          backgroundColor: Colors.redAccent,
+        ),
+      );
+    }
+    // Jika sukses, main.dart akan otomatis mengarahkan ke Home
   }
 }
