@@ -1,24 +1,63 @@
 import 'package:flutter/material.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:provider/provider.dart';
-import 'app.dart';
+import 'firebase_options.dart'; // File konfigurasi otomatis
+
+// Import Provider
 import 'providers/auth_provider.dart';
 import 'providers/court_provider.dart';
 import 'providers/booking_provider.dart';
-import 'package:intl/date_symbol_data_local.dart';
+
+// Import Screen
+import 'screens/auth/login_screen.dart';
+import 'screens/auth/register_screen.dart';
+import 'screens/home/home_screen.dart'; // Atau MainScreen jika ada navbar
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-// Inisialisasi Firebase kalau dipakai
-// await Firebase.initializeApp();
+  
+  // INI BAGIAN PENTING: Menyalakan Firebase
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
 
-  await initializeDateFormatting('id_ID', null);
+  runApp(const SportifyApp());
+}
 
-  runApp(MultiProvider(
-    providers: [
-      ChangeNotifierProvider(create: (_) => AuthProvider()),
-      ChangeNotifierProvider(create: (_) => CourtProvider()),
-      ChangeNotifierProvider(create: (_) => BookingProvider()),
-    ],
-    child: const SportifyApp(),
-  ));
+class SportifyApp extends StatelessWidget {
+  const SportifyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MultiProvider(
+      // Daftarkan semua provider Anda di sini
+      providers: [
+        ChangeNotifierProvider(create: (_) => AuthProvider()),
+        ChangeNotifierProvider(create: (_) => CourtProvider()),
+        ChangeNotifierProvider(create: (_) => BookingProvider()),
+      ],
+      child: MaterialApp(
+        title: 'Sportify',
+        debugShowCheckedModeBanner: false,
+        theme: ThemeData(
+          primarySwatch: Colors.green,
+          scaffoldBackgroundColor: Colors.grey[50],
+        ),
+        // Cek status login (akan kita buat logicnya di AuthProvider)
+        home: Consumer<AuthProvider>(
+          builder: (context, auth, _) {
+            if (auth.isLoggedIn) {
+              return const HomeScreen(); 
+            }
+            return const LoginScreen();
+          },
+        ),
+        routes: {
+          '/login': (context) => const LoginScreen(),
+          '/register': (context) => const RegisterScreen(),
+          '/home': (context) => const HomeScreen(),
+        },
+      ),
+    );
+  }
 }
